Refactor async non-GPU store into dedicated AsyncDataTransferContext
